use theme 'Sunburst' for Markdown series
So that every syntax is shown for Markdown and Markdown Extended. Note
that the `Sunburst.tmTheme` does not actually exist; `Packages` folder
does not even contain the folder `Color Scheme - Default/`. However,
the path is hardcoded in program somewhere so that when we set the
`color_scheme` as the (virtual) path, the program will search for
the actual theme file, possibly extract them from somewhere, cache
them in `Cache` folder, and provide them to user.